Become compatible with python-ldap 3
With python-ldap 3 all attribute values are stored as "bytes" to be
complaint to the RFC. This commit makes the necessary changes such
that attribute values are stored as bytes internally. Also the
set_directory method is extended with a check to refuse attribute
values that are not of type "bytes".